Fixed some files
