+++ spec/features/questions_spec.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 require 'rails_helper'
+
 feature 'Create question', '
   In order to get answer from community
   As an authenticated user
@@ -17,6 +18,8 @@ feature 'Create question', '
     fill_in :question_body, with: 'text text'
     click_on 'Create'
 
+    expect(page).to have_content 'Test question'
+    expect(page).to have_content 'text text'
     expect(page).to have_content 'Your question successfully created.'
   end
 
@@ -34,7 +37,7 @@ feature 'Browse list of questions', '
 ' do
 
   given(:user) { create(:user) }
-  given!(:questions_list) { create_list(:question, 3) }
+  given!(:questions_list) { create_pair(:question) }
 
   scenario 'Authenticated user can browse list of questions' do
     sign_in(user)
@@ -46,48 +49,20 @@ feature 'Browse list of questions', '
   end
 end
 
-feature 'User creates answer on question page', '
-  As an aunthenticated user
-  I want to be able to create answer on question page
-' do
-
-  given(:user) { create(:user) }
-  given!(:question) { create(:question) }
-
-  scenario 'Authenticated user can create answer on question page' do
-    sign_in(user)
-    visit question_path(id: question.id)
-
-    fill_in :answer_body, with: 'text text'
-    click_on 'Answer'
-  end
-
-  scenario 'Non-authenticated user tries to add answer' do
-    visit root_path
-    click_on question.title
-
-    expect(page).to have_content 'You need to sign in or sign up before continuing.'
-  end
-end
-
-feature 'User browses question and answers', '
+feature 'User browses question', '
   As an aunthenticated user
   I want to be able to browse question and answers
 ' do
 
   given(:user) { create(:user) }
   given(:question) { create(:question) }
-  given!(:answers_list) { create_list(:answer, 3, question: question) }
 
-  scenario 'Authenticated user can browse question and answers' do
+  scenario 'Authenticated user can browse question' do
     sign_in(user)
     visit question_path(id: question.id)
 
     expect(page).to have_content(question.title)
     expect(page).to have_content(question.body)
-    answers_list.each do |answer|
-      expect(page).to have_content(answer.body)
-    end
   end
 end
 
@@ -113,38 +88,6 @@ feature 'User deletes his question', '
     sign_in(user)
     visit question_path(id: another_question.id)
 
-    expect(page).not_to have_content('Delete answer')
-  end
-end
-
-feature 'User deletes his answer', '
-  As an aunthenticated user
-  I want to be able to delete my answer
-' do
-
-  given(:user) { create(:user) }
-  given(:question) { create(:question) }
-  given!(:user_answer) { create(:answer, question: question, user: user) }
-  given!(:another_answer) { create(:answer, question: question) }
-
-  before do
-    sign_in(user)
-    visit question_path(id: question.id)
-  end
-
-  scenario 'Authenticated user can delete his answer' do
-    within "#answer_#{user_answer.id}" do
-      click_on 'Delete answer'
-    end
-
-    expect(current_path).to eq question_path(question)
-    expect(page).not_to have_content(user_answer.body)
-  end
-
-  scenario "Authenticated user can't delete someone else's answer" do
-    within "#answer_#{another_answer.id}" do
-      expect(page).not_to have_content('Delete answer')
-    end
-    expect(page).to have_content(another_answer.body)
+    expect(page).not_to have_content('Delete question')
   end
 end
